requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.4.1+cu121
torchvision==0.19.1+cu121
torchaudio==2.4.1+cu121

sentence-transformers==2.5.1
qdrant-client==1.7.3
fastapi==0.115.0
uvicorn==0.30.1
transformers==4.44.2
accelerate
python-dotenv==1.0.1
pydantic==2.9.0
qdrant-client==1.7.3
numpy
requests
apscheduler
huggingface-hub==0.23.4
